added seed file and instruction and some error handling

+++ src/controller/usercontroller.ts
@@ -1,10 +1,10 @@
 import type { Request, Response } from "express";
 import bcrypt from "bcryptjs";
 import {users} from "../models/userSchema.js";
-getUsers
+
 export async function getUsers(req:Request,res:Response) {
   const userss = await users.find();
-  res.json(users);
+  res.status(201).json(userss);
 };
 
 
@@ -15,6 +15,9 @@ export async function createUser(req:Request,res:Response) {
 if(password.length < 8){
   return res.status(400).json({message:"Password must be at least 8 characters long"});
 }
+    if (!name || !email || !password) {
+      return res.status(400).json({ message: "Name, email, and password are required" });
+    }
 
 
     const hashedPassword = await bcrypt.hash(password, 10);
@@ -29,7 +32,9 @@ if(password.length < 8){
 
 export async function RemoveUsers(req:Request,res:Response) {
   const {id} = req.body
-
+ if (!id){
+   return res.status(405).json({ message: "Error Deleting user could not find id" });
+ }
   try {
     const finduserbyidremove = await users.findByIdAndDelete(id)
      res.status(200).json({message:"User has been deleted"});
@@ -48,10 +53,17 @@ export async function findandubdate(req:Request,Res:Response) {
 
   const update = req.body
    const {id} = req.params
-  update.password = await bcrypt.hash(update.password,10)
 
+   if(update.password){
+     update.password = await bcrypt.hash(update.password,10)
+   }
 
- 
+if(!id){
+  return Res.status(405).json({ message: "Error Updating user could not find id" });
+}
+if (!update || Object.keys(update).length === 0) {
+  return Res.status(400).json({ message: "Error updating user: no update data provided" });
+}
 
   try {
